+++ backend/src/main/java/org/codecritique/thrifty/controller/LabelsController.java
@@ -6,6 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
@@ -17,6 +21,7 @@ public class LabelsController extends BaseController {
     @Autowired
     private LabelDao dao;
 
+    @PreAuthorize("hasAuthority(#label.accountId)")
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Resource> createLabel(@RequestBody Label label) {
         dao.store(label);
@@ -24,6 +29,7 @@ public class LabelsController extends BaseController {
         return ResponseEntity.created(location).build();
     }
 
+    @PreAuthorize("hasAuthority(#label.accountId)")
     @PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Resource> updateLabel(@RequestBody Label label) {
         dao.updateLabel(label);
@@ -31,20 +37,25 @@ public class LabelsController extends BaseController {
     }
 
     @DeleteMapping(path = "{id}")
-    public ResponseEntity<Resource> removeLabel(@PathVariable long id) {
+    public ResponseEntity<Resource> removeLabel(@PathVariable long id, @AuthenticationPrincipal UserDetails userDetails) {
         Label label = dao.getLabel(id);
         if (label == null)
             return ResponseEntity.notFound().build();
-        dao.removeLabel(id);
-        return ResponseEntity.ok().build();
+        if (hasAuthority(userDetails, label)) {
+            dao.removeLabel(id);
+            return ResponseEntity.ok().build();
+        }
+        throw new AccessDeniedException("Access is denied");
     }
 
     @GetMapping(path = "{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Label> getLabel(@PathVariable long id) {
+    public ResponseEntity<Label> getLabel(@PathVariable long id, @AuthenticationPrincipal UserDetails userDetails) {
         Label label = dao.getLabel(id);
         if (label == null)
             return ResponseEntity.notFound().build();
-        return ResponseEntity.ok(label);
+        if (hasAuthority(userDetails, label))
+            return ResponseEntity.ok(label);
+        throw new AccessDeniedException("Access is denied");
     }
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
